pick up function

+++ BL/DroneInParcel.cs
@@ -9,7 +9,7 @@ namespace IBL.BO
     public class DroneInParcel
     {
         public int Id { get; set; }
-        public int Battery { get; set; }
+        public double Battery { get; set; }
         public Location CurrentLocation { get; set; }
 
     }
